+++ src/main/java/de/dealog/msg/converter/UnsupportedConversionException.java
@@ -1,0 +1,9 @@
+package de.dealog.msg.converter;
+
+public class UnsupportedConversionException extends RuntimeException {
+
+    public UnsupportedConversionException(final String msg) {
+        super(msg);
+    }
+
+}