+++ src/main/java/de/dealog/msg/persistence/model/Region.java
@@ -1,0 +1,9 @@
+package de.dealog.msg.persistence.model;
+
+public interface Region {
+    String getCode();
+
+    String getName();
+
+    String getDescription();
+}