+++ src/main/java/de/dealog/msg/persistence/repository/MessageRepository.java
@@ -6,9 +6,6 @@ import io.quarkus.hibernate.orm.panache.PanacheRepository;
 
 import javax.enterprise.context.ApplicationScoped;
 
-/**
- * Implementation of a {@link MessageEntity} repository
- */
 @ApplicationScoped
 public class MessageRepository implements PanacheRepository<MessageEntity> {