+++ src/main/java/de/dealog/msg/rest/model/RegionTypeRest.java
@@ -1,0 +1,36 @@
+package de.dealog.msg.rest.model;
+
+import lombok.Getter;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+
+@Getter
+public enum RegionTypeRest {
+
+    COUNTRY("Bundesrepublik"),
+    STATE("Land", "Freie Hansestadt", "Freie und Hansestadt", "Freistaat"),
+    COUNTY("Regierungsbezirk"),
+    DISTRICT("Kreis", "Kreisfreie Stadt", "Landkreis", "Stadtkreis"),
+    MUNICIPALITY("Gemeinde", "Gemeindefreies Gebiet", "Stadt"),
+    UNKNOWN;
+
+    private final List<String> identifiers;
+
+    RegionTypeRest(final String... identifiers) {
+        this.identifiers = Arrays.asList(identifiers);
+    }
+
+    /**
+     * Find a region type by an identifer.
+     *
+     * @param identifier the identifier as string
+     * @return an {@link Optional} containing the found {@link RegionTypeRest}
+     */
+    public static Optional<RegionTypeRest> findByIdentifier(final String identifier) {
+        return Arrays.stream(RegionTypeRest.values())
+        .filter(rt -> rt.getIdentifiers().contains(identifier))
+        .findFirst();
+    }
+}