+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 1.0.0 (under development)
 This is the initial version setting up the base system
 
+- `NEW` Add region code to message and ability to find messages by region code
+- `NEW` Add region resource to find regions by region code or location
+- `NEW` Add REST method to return a region hierachy by ars or location
+- `NEW` Basic region model
 - 'CNG' Change to Java 11
 - `NEW` Add Message resource to find a message by its identifier 
 - `CNG` Resource path 'api/message' is now 'api/messages'